Initial edits to `packet.rs`, `host.rs`, `router.rs`, `routing_trie.rs`
Filled in the structs and main functions of each of the files above. There might be some extra helper functions in `router.rs` that aren't being used that are in relation with `routing_trie.rs`, but these can be deleted later. Not sure if the implementation of `routing_trie.rs` is correct, but the logic behind the usage of the routing tree in `router.rs` should be correct, but this needs to be checked.