Added TODOs in `network_interface.rs` and `host.rs`, and restructured both files generally
The restructuring revolves around having the UDP socket be located in each interface. Each interface will have:
- udp socket
- listening thread
- separate function for sending data
- pointer to parent device

When receiving data, the interface should process that by sending it up to the parent device to see what should be done with the packet. The listening thread should therefore consistently be running in a loop. The socket should be locked and unlocked during specific send requests (REVIEW THIS).

The `host.rs` changes revolve around the changes to the network interface implementation as well, as there is no need for udp sockets in that file. Similar changes should be made to `router.rs` as well.