Modified most of the REPL and channel logistics to handle everything pre packet processing
Changed a lot of the structs so that there is always access to any information that is needed. Established channel connections (only for `host.rs` right now, routers are a little different since there is more than one thread being spawned per interface...)
Still need to work on actual processing of the data packets and IP forwarding.